+++ union/lib/union.dart
@@ -1,5 +1,6 @@
 library union;
 
 export 'src/annotations.dart' show Union;
+export 'src/cases.dart';
 export 'src/errors.dart';
-export 'src/models.dart';
+export 'src/params.dart';