+++ union_generator/lib/src/common/declarations/union_declaration.dart
@@ -1,7 +1,12 @@
 import 'package:analyzer/dart/element/element.dart';
+import 'package:union/union.dart';
 
 abstract class UnionDeclaration<T extends Element> {
-  const UnionDeclaration(this.target);
+  const UnionDeclaration(
+    this.target,
+    this.paramsType,
+  );
 
   final T target;
+  final UnionParamsType paramsType;
 }