+++ union_generator/lib/src/parsers/class_types_union_parser.dart
@@ -19,9 +19,10 @@ class ClassTypesUnionParser
 
   ClassTypesUnionDeclaration _createDeclaration(Element element, ConstantReader annotation) {
     final targetElement = getTargetElement(element, annotation);
+    final paramsType = getParamsType(annotation);
     final types =
         annotation.read('types').setValue.map((dartObject) => dartObject.toTypeValue()!).toSet();
 
-    return ClassTypesUnionDeclaration(targetElement, types);
+    return ClassTypesUnionDeclaration(targetElement, paramsType, types);
   }
 }