+++ union_generator/lib/src/parsers/enum_union_cases_parser.dart
@@ -28,9 +28,10 @@ class EnumCasesUnionParser
 
   EnumCasesUnionDeclaration _createDeclaration(Element element, ConstantReader annotation) {
     final targetElement = getTargetElement(element, annotation);
+    final paramsType = getParamsType(annotation);
     final cases = _getEnumCases(targetElement, annotation);
 
-    return EnumCasesUnionDeclaration(targetElement, cases);
+    return EnumCasesUnionDeclaration(targetElement, paramsType, cases);
   }
 
   Set<EnumCaseDeclaration> _getEnumCases(EnumElement targetElement, ConstantReader annotation) {